move mock entry point
