+++ src/promptflow/tests/executor/unittests/_core/test_tools_manager.py
@@ -20,6 +20,16 @@ from promptflow.contracts.tool import Tool, ToolType
 from promptflow.exceptions import UserErrorException
 
 
+@pytest.fixture
+def mock_entry_point():
+    from ...package_tools.custom_llm_tool_multi_inputs_without_index.list import list_package_tools
+    entry_point = MagicMock()
+    entry_point.load.return_value = list_package_tools
+    entry_point.dist.metadata.return_value = "TestCustomLLMTool"
+    entry_point.dist.version.return_value = "0.0.1"
+    return entry_point
+
+
 @pytest.mark.unittest
 class TestToolLoader:
     def test_load_tool_for_node_with_invalid_node(self):
